requirements.txt

phi==0.6.7
openai
yfinance
python-dotenv
